# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Baseball Lineup Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tournaments:
    get:
      operationId: listTournaments
      tags: [tournaments]
      summary: List the coach's tournaments (most recent first)
      responses:
        "200":
          description: Tournaments with rolling stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TournamentSummary"
    post:
      operationId: createTournament
      tags: [tournaments]
      summary: Create a tournament
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTournamentBody"
      responses:
        "201":
          description: Created tournament
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
...
  /tournaments/{id}/network:
    get:
      operationId: getTournamentNetwork
      tags: [tournaments]
      summary: Get tournament network state + join suggestions
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Network state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TournamentNetworkResponse"
        "404": { description: Tournament not found }
    patch:
      operationId: updateTournamentNetworkMembership
      tags: [tournaments]
      summary: Toggle visibility / score-sharing for this tournament's membership
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTournamentNetworkMembershipBody"
      responses:
        "200":
          description: Updated membership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TournamentNetworkMember"
        "404": { description: Tournament not in any network }
  /tournaments/{id}/network/join:
    post:
      operationId: joinTournamentNetwork
      tags: [tournaments]
      summary: Join (or create) the network for this tournament's fingerprint
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "201":
          description: Joined
          content:
            application/json:
              schema:
                type: object
                properties:
                  network:
                    $ref: "#/components/schemas/TournamentNetwork"
                  membership:
                    $ref: "#/components/schemas/TournamentNetworkMember"
                required: [network, membership]
        "200":
          description: Already a member (idempotent)
          content:
            application/json:
              schema:
                type: object
                properties:
                  network:
                    $ref: "#/components/schemas/TournamentNetwork"
                  membership:
                    $ref: "#/components/schemas/TournamentNetworkMember"
                required: [network, membership]
        "400": { description: Tournament has no fingerprint }
        "409": { description: Tournament already in a different network }
  /tournaments/{id}/network/leave:
    post:
      operationId: leaveTournamentNetwork
      tags: [tournaments]
      summary: Leave the network for this tournament
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Left }
  /tournaments/{id}/network/dismiss:
    post:
      operationId: dismissTournamentNetworkSuggestions
      tags: [tournaments]
      summary: Dismiss current network join suggestions for this tournament
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Dismissed }
...
components:
  schemas:
...
    RestTier:
      type: object
      description: One row of the rest-tier ladder (pitches → required days rest).
      properties:
        maxPitches:
          type: integer
          minimum: 0
          description: Inclusive upper bound of pitch totals this tier covers.
        daysRest:
          type: integer
          minimum: 0
          description: Required calendar days of rest before pitching again.
      required:
        - maxPitches
        - daysRest
...
    Tournament:
      type: object
      properties:
        rowVersion:
          type: integer
          description: Optimistic-concurrency version, incremented on every update. Send back in If-Match on PATCH/DELETE to detect conflicts.
        id:
          type: integer
        name:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        dailyPitchMax:
          type: ["integer", "null"]
          description: Per-pitcher daily cap for this tournament. Null = inherit team default.
        tournamentPitchMax:
          type: ["integer", "null"]
          description: Per-pitcher cap across the whole tournament. Null = inherit team default.
        restTiers:
          oneOf:
            - type: "null"
            - type: array
              items:
                $ref: "#/components/schemas/RestTier"
          description: Rest tiers for this tournament. Null = inherit team default.
        poolPlayNoNewInningMinutes:
          type: ["integer", "null"]
          description: Pool-play "no new inning after N minutes" rule. Null = no rule.
        poolPlayHardStopMinutes:
          type: ["integer", "null"]
          description: Pool-play hard stop (minutes). Field Display surfaces warnings — does NOT auto-finalize. Null = no rule.
        bracketNoNewInningMinutes:
          type: ["integer", "null"]
          description: Bracket-play "no new inning after N minutes" rule. Null = no rule.
        bracketHardStopMinutes:
          type: ["integer", "null"]
          description: Bracket-play hard stop (minutes). Null = no rule.
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - startDate
        - endDate
        - createdAt
    TournamentSummary:
      allOf:
        - $ref: "#/components/schemas/Tournament"
        - type: object
          properties:
            gameCount:
              type: integer
            pitchersUsed:
              type: integer
              description: Distinct player count with at least one recorded pitch in this tournament
            totalPitches:
              type: integer
              description: Sum of pitches across all games in this tournament
          required:
            - gameCount
            - pitchersUsed
            - totalPitches
...
    CreateTournamentBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        dailyPitchMax:
          type: ["integer", "null"]
        tournamentPitchMax:
          type: ["integer", "null"]
        restTiers:
          oneOf:
            - type: "null"
            - type: array
              items:
                $ref: "#/components/schemas/RestTier"
        poolPlayNoNewInningMinutes:
          type: ["integer", "null"]
          minimum: 1
          maximum: 360
        poolPlayHardStopMinutes:
          type: ["integer", "null"]
          minimum: 1
          maximum: 360
        bracketNoNewInningMinutes:
          type: ["integer", "null"]
          minimum: 1
          maximum: 360
        bracketHardStopMinutes:
          type: ["integer", "null"]
          minimum: 1
          maximum: 360
      required:
        - name
        - startDate
        - endDate
...
    TournamentNetwork:
      type: object
      description: Cross-coach network row keyed on tournament fingerprint.
      properties:
        id: { type: integer }
        fingerprint: { type: string }
        displayName: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, fingerprint, displayName, createdAt]
    TournamentNetworkMember:
      type: object
      description: A tournament's membership in a network. One row per tournament.
      properties:
        id: { type: integer }
        networkId: { type: integer }
        tournamentId: { type: integer }
        userId: { type: string }
        visible:
          type: boolean
          description: Coach opted to show team name to other members. Default false.
        shareScores:
          type: boolean
          description: Coach is sharing scored games with the network. Default true.
        joinedAt: { type: string, format: date-time }
      required: [id, networkId, tournamentId, userId, visible, shareScores, joinedAt]
    TournamentNetworkMemberSummary:
      type: object
      description: Public view of a network member surfaced to other coaches.
      properties:
        tournamentId: { type: integer }
        visible: { type: boolean }
        shareScores: { type: boolean }
        teamName:
          type: ["string", "null"]
          description: Null when the member has not opted in to be visible.
        isYou: { type: boolean }
      required: [tournamentId, visible, shareScores, isYou]
    TournamentNetworkSuggestion:
      type: object
      description: A matching tournament owned by another coach who hasn't joined yet.
      properties:
        tournamentId: { type: integer }
        name: { type: string }
        teamName: { type: ["string", "null"] }
        networkId: { type: ["integer", "null"] }
      required: [tournamentId, name]
    TournamentNetworkResponse:
      type: object
      properties:
        network:
          oneOf:
            - $ref: "#/components/schemas/TournamentNetwork"
            - type: "null"
        membership:
          oneOf:
            - $ref: "#/components/schemas/TournamentNetworkMember"
            - type: "null"
        members:
          type: array
          items:
            $ref: "#/components/schemas/TournamentNetworkMemberSummary"
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/TournamentNetworkSuggestion"
      required: [members, suggestions]
    UpdateTournamentNetworkMembershipBody:
      type: object
      properties:
        visible: { type: boolean }
        shareScores: { type: boolean }